+++ src/lib.rs
@@ -18,6 +18,7 @@ pub mod product;
 pub mod reboot;
 pub mod resume;
 pub mod stats_ini;
+pub mod system_exec;
 pub mod ui;
 pub mod uninstall;
 pub mod vendor;